% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccuModel.R
\name{AccuModel-deprecated}
\alias{AccuModel-deprecated}
\title{AccuModel}
\description{
AccuModel
}
\seealso{
<<<<<<< HEAD
[TestDimorph-deprecated()]
=======
\code{\link[=TestDimorph-deprecated]{TestDimorph-deprecated()}}
>>>>>>> 6dea5f88f050c15e544b721a245e69cf92b48861
}
\keyword{internal}
