% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{anova_es}
\alias{anova_es}
\title{anova_es}
\usage{
anova_es(x, es = es, digits = digits)
}
\arguments{
\item{x}{ANOVA model}

\item{es}{Logical; if TRUE effect size is included in the output , Default:
FALSE}

\item{digits}{Number of significant digits, Default: 4}
}
\description{
anova_es
}
\keyword{internal}
