% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA.R
\name{anova_main_II}
\alias{anova_main_II}
\title{ANOVA type II for main effects}
\usage{
anova_main_II(x, es_aov, digits, CI, lower.tail)
}
\arguments{
\item{x}{A data frame containing summary statistics.}

\item{digits}{Number of significant digits, Default: 4}

\item{CI}{confidence interval coverage takes value from 0 to 1, Default: 0.95.}

\item{lower.tail}{Logical; if TRUE probabilities are `P[X <= x]`,
otherwise, `P[X > x]`., Default: FALSE}
}
\description{
ANOVA type II for main effects
}
\keyword{internal}
