% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Australia}
\alias{Australia}
\title{Australia}
\format{
A data frame with 94 rows and 9 variables:
 \describe{
\item{Pop}{(Region) ("B1" = Southwest Australia, "B19" = Northeast Australia),
 see Gilligan and Bulbeck (2007)}
\item{Sex}{Sex coded as "F" or "M"}
\item{Weight.kg}{body weight in kilograms}
\item{Stature.mm}{Standing height in millimeters}
\item{Hum.Lgth}{Humeral length in millimeters}
\item{Rad.Lgth}{Radius length in millimeters}
\item{Fem.Lgth}{Femoral length in millimeters}
\item{Tib.Lgth}{Tibial length in millimeters}
\item{Bi.illiac}{Bi-illiac breadth in millimeters}
  }
}
\usage{
Australia
}
\description{
Raw data from Joseph Birdsell's 1938 survey of Australian Aborigines
(available from \url{http://www.peterbrown-palaeoanthropology.net/resource.html})
 The data is from two regions (B1 and B19), see Gilligan and Bulbeck (2007)
 for a map of the regions.
}
\references{
Gilligan, I., & Bulbeck, D. (2007). Environment and morphology in Australian
Aborigines: A re‐analysis of the Birdsell database. American Journal of Physical
Anthropology, 134(1), 75-91.
}
\keyword{datasets}
