% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baboon.parms_df}
\alias{baboon.parms_df}
\title{data frame format for the baboon.parms_df for multivariate analysis}
\format{
A data frame with 12 rows and 8 variables \describe{
\item{Trait}{Apolipoprotein B and LDL on two diets} \item{Sub}{Sub-species or hybrid}
\item{M.mu}{Means of LDL and apo B in different sub-species for males}
\item{F.mu}{Means of LDL and apo B in different sub-species for females}
\item{m}{Male sample sizes} \item{f}{Female sample sizes}
\item{M.sdev}{Standard deviations for males} \item{F.sdev}{Standard
deviations for females}

 }
}
\usage{
baboon.parms_df
}
\description{
A dataset containing summary statistics for low density lipoprotein (LDL) and
apolipoprotein B (apo B) levels in 604 baboons measured on two different diets:
a basal diet and a high cholesterol, saturated fat diet. The baboons were
classified into one of two subspecies and a hybrid of the two subspecies
(Papio hamadryas anubis, P.h. cynocephalus, or hybrid). Each animal was
measured on each of the two diets.
}
\note{
The baboon data collection were supported by NIH grant HL28972 and
NIH contract HV53030 to the Southwest Foundation for Biomedical Research
(Now: Texas Biomedical Research Institute), and funds from the Southwest
Foundation for Biomedical Research
}
\references{
Konigsberg LW (1991). An historical note on the t-test for differences in
sexual dimorphism between populations. American journal of physical
anthropology, 84(1), 93–96.
}
\keyword{datasets}
