% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baboon.parms_list}
\alias{baboon.parms_list}
\title{List format for the baboon.parms_df for multivariate analysis}
\format{
A list of 7 matrices.
\describe{ \item{R.res}{pooled within group
correlation matrix}
\item{M.mu}{Means of lipoproteins in different species
for males}
\item{F.mu}{Means of lipoproteins in different species for
females}
\item{m}{Male sample sizes} \item{f}{Female sample sizes}
\item{M.sdev}{Standard deviations for males} \item{F.sdev}{Standard
deviations for females} }
}
\usage{
baboon.parms_list
}
\description{
List format for the baboon.parms_df for multivariate analysis
}
\keyword{datasets}
