% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cremains_measurements}
\alias{Cremains_measurements}
\title{Measurements from calcined postcranial materials.}
\format{
A data frame with 22 rows and 8 variables:
\describe{
\item{Trait}{Measured feature}
\item{M.mu}{Means of males}
\item{F.mu}{Means of females}
\item{m}{Male sample sizes} \item{f}{Female sample sizes}
\item{M.sdev}{Standard deviations for males} \item{F.sdev}{Standard
deviations for females}\item{D}{published value for Chakraborty and Majumder's
(1982) measure of sexual dimorphism.}
}
}
\usage{
Cremains_measurements
}
\description{
Part of Table 3 from Cavazzuti et al. (2019).
}
\references{
Cavazzuti, Claudio, et al. (2019) "Towards a new osteometric method for sexing
ancient cremated human remains. Analysis of Late Bronze Age and Iron Age samples
from Italy with gendered grave goods." PloS one 14.1: e0209423.

Chakraborty, R., & Majumder, P. P. (1982). On Bennett's measure of sex
dimorphism. American journal of physical anthropology, 59(3), 295-298.
}
\keyword{datasets}
