% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FT}
\alias{FT}
\title{Heuristic data}
\format{
A data frame with 24 rows and 3 variables:
 \describe{
\item{Sex}{'M' for male and 'F' for female}
\item{Pop}{Populations' names}
\item{x}{Dependent variable}
  }
}
\usage{
FT
}
\description{
Heuristic data from Fidler and Thompson (2001)
}
\references{
Fidler, Fiona, and Bruce Thompson. "Computing correct confidence intervals
for ANOVA fixed-and random-effects effect sizes." Educational and
Psychological Measurement 61.4 (2001): 575-604.
}
\keyword{datasets}
