% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Howells_R}
\alias{Howells_R}
\title{Pooled within group correlation matrix for Howells' data}
\format{
A 8*8 numerical matrix
}
\usage{
Howells_R
}
\description{
Pooled within group correlation matrix for Howells' data
}
\keyword{datasets}
