% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Howells_summary}
\alias{Howells_summary}
\title{Summary of the Howells' craniometric data}
\format{
A data frame with 32 rows and 8 variables:
\describe{
\item{Trait}{Measured feature} \item{Pop}{Population name}
\item{M.mu}{Means of males}
\item{F.mu}{Means of females}
\item{m}{Male sample sizes} \item{f}{Female sample sizes}
\item{M.sdev}{Standard deviations for males} \item{F.sdev}{Standard
deviations for females}
}
}
\usage{
Howells_summary
}
\description{
Summary statistics of the the Howells' data subset.
}
\references{
\link{Howells}
}
\keyword{datasets}
