% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Howells_summary_list}
\alias{Howells_summary_list}
\title{List format of \link{Howells_summary} for multivariate analysis}
\format{
A list of 5 matrices (R.res, M.mu, F.mu, M.sdev, and F.sdev) and
two vectors (m and f) with structure similar to \link{baboon.parms_list}
}
\usage{
Howells_summary_list
}
\description{
List format of \link{Howells_summary} for multivariate analysis
}
\keyword{datasets}
