% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Howells_V}
\alias{Howells_V}
\title{Pooled within-group variance-covariance matrix for Howells' data}
\format{
A 8*8 numerical matrix
}
\usage{
Howells_V
}
\description{
Pooled within-group variance-covariance matrix for Howells' data
}
\keyword{datasets}
