% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_D.R
\name{index_D}
\alias{index_D}
\title{Dissimilarity index}
\usage{
index_D(
  x,
  plot = FALSE,
  fill = c("female", "male", "both"),
  r = NULL,
  Trait = 1
)
}
\arguments{
\item{x}{A data frame containing summary statistics.}

\item{plot}{Logical; if true a plot of densities is returned, Default:
FALSE}

\item{fill}{Specify which sex to be filled with color in the plot; either
"male" in blue color, "female" in pink color or "both", Default: 'female'}

\item{r}{Number of row for a desired trait, Default: NULL}

\item{Trait}{Number of column containing trait names as factors, Default: 1}
}
\value{
A tibble and a graphical representation of the chosen traits and
their corresponding dissimilarity indices
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\keyword{internal}
