% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MANOVA.R
\name{manova_III}
\alias{manova_III}
\title{MANOVA type III}
\usage{
manova_III(x, es_manova, test, digits, CI, lower.tail)
}
\arguments{
\item{x}{Data frame or list containing summary statistics for
multiple parameters measured in both sexes in two or more populations.}

\item{es_manova}{effect size either ,"eta" for eta squared, or "none"for
not reporting an effect size, Default:"none".}

\item{digits}{Number of significant digits, Default: 4}

\item{CI}{confidence interval coverage takes value from 0 to 1, Default: 0.95.}

\item{lower.tail}{Logical; if TRUE probabilities are `P[X <= x]`,
otherwise, `P[X > x]`., Default: FALSE}
}
\description{
MANOVA type III
}
\keyword{internal}
