% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NHANES_1999}
\alias{NHANES_1999}
\title{NHANES 1999}
\format{
A data frame with 1430 rows and 5 variables:
 \describe{
\item{Sex}{(RIAGENDR) Sex coded as "F" or "M"}
\item{Pop}{(RIDRETH1) Self-reported race, coded as "Black" = Non-Hispanic Black,
"Mex.Am" = Mexican American, or "White" = Non-Hispanic White}
\item{BMXWT}{Body weight in kilograms}
\item{BMXHT}{Standing height in centimeters}
\item{BMXARML}{Upper arm length in centimeters}
  }
}
\usage{
NHANES_1999
}
\description{
Raw data from 1999-2000 NHANES (National Health and Nutrition
Examination Survey). Centers for Disease Control and Prevention (CDC). National
Center for Health Statistics (NCHS). National Health and Nutrition Examination
Survey Data. Hyattsville, MD: U.S. Department of Health and Human Services, Centers
for Disease Control and Prevention, 2020, \url{https://www.cdc.gov/nchs/nhanes/index.htm}
}
\note{
This is not the complete dataset. It is selected so that age in years is
greater than or equal to 20 and less than or equal to 40
}
\keyword{datasets}
