% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pooled_cov}
\alias{pooled_cov}
\title{pooled_cov}
\usage{
pooled_cov(x, ina)
}
\arguments{
\item{x}{A matrix with continuous data}

\item{ina}{A numerical vector indicating the groups}
}
\description{
Pooled covariance matrix
}
\keyword{internal}
