% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{R}
\alias{R}
\title{Pooled within group correlation matrix for baboon data}
\format{
A 4*4 numerical matrix
}
\usage{
R
}
\description{
Pooled within group correlation matrix for baboon data
}
\keyword{datasets}
