% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rown_col}
\alias{rown_col}
\title{rown_col}
\usage{
rown_col(x, var)
}
\arguments{
\item{x}{dataframe with rownames}

\item{var}{new column name}
}
\description{
rown_col
}
\details{
convert rownames to column
}
\keyword{internal}
