% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SMO}
\alias{SMO}
\title{Hypothetical set of unbalanced data}
\format{
A data frame with 11 rows and 3 variables:
 \describe{
\item{Sex}{'M' for male and 'F' for female}
\item{Pop}{Populations' names}
\item{x}{Dependent variable}
  }
}
\usage{
SMO
}
\description{
Example data set from Shaw and Mitchell-Olds (1993)
}
\references{
Shaw, Ruth G., and Thomas Mitchell-Olds. "ANOVA for unbalanced data: an overview.
" Ecology 74.6 (1993): 1638-1645.
}
\keyword{datasets}
