% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{t_test}
\alias{t_test}
\title{t-greene test}
\usage{
t_test(
  m,
  f,
  m2,
  f2,
  M.mu,
  F.mu,
  M.mu2,
  F.mu2,
  M.sdev,
  F.sdev,
  M.sdev2,
  F.sdev2,
  padjust = padjust,
  N,
  digits,
  CI,
  alternative,
  es
)
}
\arguments{
\item{m}{Number of male sample size in the first population, Default: NULL}

\item{f}{Number of female sample size in the first population, Default:
NULL}

\item{m2}{Number of male sample size in the second population, Default:
NULL}

\item{f2}{Number of female sample size in the second population, Default:
NULL}

\item{M.mu}{Means for males in the first population, Default: NULL}

\item{F.mu}{Means for females in the first population, Default: NULL}

\item{M.mu2}{Means for males in the second population, Default: NULL}

\item{F.mu2}{Means for females in the second population, Default: NULL}

\item{M.sdev}{Standard deviation for males in the first population,
Default: NULL}

\item{F.sdev}{Standard deviation for females in the first population,
Default: NULL}

\item{M.sdev2}{Standard deviation for males in the second population,
Default: NULL}

\item{F.sdev2}{Standard deviation for females in the second population,
Default: NULL}

\item{padjust}{Method of p.value adjustment for multiple comparisons
following \link[stats]{p.adjust.methods}  Default: "none".}

\item{N}{Number of pairwise comparisons for [p.adjust.methods], if left
`NULL` it will follow the formula `n(n 1)/2` where `n` is the number of
populations , Default: NULL}

\item{digits}{Number of significant digits, Default: 4}

\item{CI}{confidence interval coverage takes value from 0 to 1, Default: 0.95.}

\item{alternative}{a character string specifying the alternative
hypothesis, must be one of "two.sided", "greater" or "less".}

\item{es}{Type of effect size either "d" for Cohen's d,"g" for Hedge's g
or "none" , Default:"none".}
}
\description{
t-greene test
}
\keyword{internal}
