% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestDimorph-deprecated.R, R/pMatrix.R
\name{TestDimorph-deprecated}
\alias{TestDimorph-deprecated}
\alias{pMat}
\title{Deprecated functions in package \pkg{TestDimorph}.}
\usage{
pMat(
  x = NULL,
  Pop = 1,
  plot = FALSE,
  padjust = "none",
  alternative = "two.sided",
  ...
)
}
\description{
The functions listed below are deprecated and will be defunct in
the near future.
}
\section{\code{pMatrix}}{

For \code{pMatrix}, use \code{\link{Tg}}.
}

\keyword{internal}
