% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tg.R
\name{Tg-deprecated}
\alias{Tg-deprecated}
\title{Tg}
\description{
Tg
}
\seealso{
<<<<<<< HEAD
[TestDimorph-deprecated()]
=======
<<<<<<< HEAD
[TestDimorph-deprecated()]
=======
\code{\link[=TestDimorph-deprecated]{TestDimorph-deprecated()}}
>>>>>>> 6dea5f88f050c15e544b721a245e69cf92b48861
>>>>>>> 3ffea2d74b01d4945f77061b801bd9783c34951e
}
\keyword{internal}
