% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_parms.R
\name{to_params}
\alias{to_params}
\title{FUNCTION_TITLE}
\usage{
to_params(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\keyword{internal}
