% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/to_symm_diag.R
\name{to_symm_diag}
\alias{to_symm_diag}
\title{Get pooled within group correlation matrix & standard deviations.}
\usage{
to_symm_diag(x)

to_symm_diag(x)
}
\arguments{
\item{x}{upper triangular matrix stored by rows.}
}
\value{
a symmetric matrix.

a symmetric matrix.
}
\description{
Get pooled within group correlation matrix & standard deviations.
}
\details{
A utility function to take an upper triangular matrix stored by
rows and convert it to a symmetric matrix.

A utility function to take an upper triangular matrix stored by
rows and convert it to a symmetric matrix.
}
\examples{
#the matrix: [,1] [,2]
#[,3] [1,] 3089 1079 785 [2,] 1079 574 351 [3,] 785 351 330 Should be
#passed as the vector: c(3089, 1079, 785, 574, 351, 350).
to_symm_diag(c(3089, 1079, 785, 574, 351, 350))

#the matrix: [,1] [,2]
#[,3] [1,] 3089 1079 785 [2,] 1079 574 351 [3,] 785 351 330 Should be
#passed as the vector: c(3089, 1079, 785, 574, 351, 350).
to_symm_diag(c(3089, 1079, 785, 574, 351, 350))

}
