% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_to_R.R
\name{V_to_R}
\alias{V_to_R}
\title{Correlation matrix extraction}
\usage{
V_to_R(V)
}
\arguments{
\item{V}{Variance-covariance matrix}
}
\value{
Correlation matrix
}
\description{
Returns a correlation matrix from a variance covariance matrix
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
