% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/van_Vark.R
\name{van_Vark}
\alias{van_Vark}
\title{FUNCTION_TITLE}
\usage{
van_Vark(x, Sex = 1, Pop = 2, N = 3, W, q = 2, plot = TRUE)
}
\arguments{
\item{x}{Data frame of raw data.}

\item{Sex}{Number of the column containing sex 'M' for male and 'F' for
female, Default: 1}

\item{Pop}{Number of the column containing populations' names, Default: 2}

\item{N}{Number of column containing sample size, Default: 3}

\item{W}{Pooled within-group variance-covariance matrix}

\item{q}{Number of canonical variates to retain for chi square test,
Default: 2}

\item{plot}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
