% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Van_vark_raw}
\alias{Van_vark_raw}
\title{Van_vark_raw}
\usage{
Van_vark_raw(x, Sex, Pop, firstX, ...)
}
\arguments{
\item{x}{Data frame of raw data.}

\item{Sex}{Number of the column containing sex 'M' for male and 'F' for
female, Default: 1}

\item{Pop}{Number of the column containing populations' names, Default: 2}

\item{firstX}{Number of column containing measured parameters (First of
multiple in case of multivariate analysis), Default: 3}

\item{...}{Additional arguments that could be passed to the test of choice}
}
\description{
runs van_vark function on raw data
}
\keyword{internal}
