% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MANOVA_stat.R
\name{Wilks}
\alias{Wilks}
\title{Wilks}
\usage{
Wilks(p, H, E, v.h, v.e, CI, lower.tail)
}
\arguments{
\item{p}{Number of dependent variables (measurements)}

\item{H}{Hypothesis SSCP matrix}

\item{E}{Error SSCP matrix}

\item{v.h}{Hypothesis degrees of freedom}

\item{v.e}{Error degrees of freedom}
}
\description{
Wilks
}
\keyword{internal}
