% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{anova_es}
\alias{anova_es}
\title{anova_es}
\usage{
anova_es(x, es_anova = es_anova, digits = digits, CI = CI)
}
\arguments{
\item{x}{ANOVA model}

\item{es_anova}{Type of effect size either "f" for f squared,"eta" for eta
squared,"omega" for omega squared or "none", Default:"none".}

\item{digits}{Number of significant digits, Default: 4}

\item{CI}{confidence interval coverage takes value from 0 to 1, Default: 0.95.}
}
\description{
anova_es
}
\keyword{internal}
