% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baboon.parms_list}
\alias{baboon.parms_list}
\title{Summary statistics of baboon data collection data frame A dataset
containing summary statistics for low density lipoprotein (LDL) and
apolipoprotein B (apo B) levels in 604 baboons measured on two different
diets: a basal diet 'chow' and a high cholesterol, saturated fat diet
'pink' (HCSF). The baboons were classified into one of three subspecies
(Papio hamadryas anubis, P.h. cynocephalus, or anubistcynocephalus hybrid).
Each animal was measured on each of the two diets.}
\format{
A list of 7 matrices.
\describe{ \item{R.res}{pooled within group
correlation matrix}
\item{M.mu}{Means of lipoproteins in different species
for males}
\item{F.mu}{Means of lipoproteins in different species for
females}
\item{m}{Male sample sizes} \item{f}{Female sample sizes}
\item{M.sdev}{Standard deviations for males} \item{F.sdev}{Standard
deviations for females} }
}
\usage{
baboon.parms_list
}
\description{
Summary statistics of baboon data collection data frame A dataset
containing summary statistics for low density lipoprotein (LDL) and
apolipoprotein B (apo B) levels in 604 baboons measured on two different
diets: a basal diet 'chow' and a high cholesterol, saturated fat diet
'pink' (HCSF). The baboons were classified into one of three subspecies
(Papio hamadryas anubis, P.h. cynocephalus, or anubistcynocephalus hybrid).
Each animal was measured on each of the two diets.
}
\note{
The baboon data collection were supported by NIH grant HL28972 and
NIH contract HV53030 to the Southwest Foundation for Biomedical Research
(Now: Texas Biomedical Research Institute), and funds from the Southwest
Foundation for Biomedical Research
#' @references

Konigsberg LW (1991). \emph{An historical note on the t-test for differences in sexual dimorphism
between populations.}. American journal of physical anthropology, 84(1), 93–96.
}
\keyword{datasets}
