% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{eff_CI}
\alias{eff_CI}
\title{eff_CI}
\usage{
eff_CI(f, CI, eff, df1, df2)
}
\arguments{
\item{f}{critical F-value}

\item{CI}{critical p-value, Default 0.05}

\item{eff}{effect size}

\item{df1}{numerator degree of freedom}

\item{df2}{denominator degree of freedom}
}
\description{
Confidence intervals for ANOVA and MANOVA effect sizes
}
\keyword{internal}
