% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sum.R
\name{extract_sum}
\alias{extract_sum}
\title{Summary Statistics Extraction}
\usage{
extract_sum(x, Sex = 1, Pop = 2, firstX = 3, test = 1, run = TRUE, ...)
}
\arguments{
\item{x}{Data frame of raw data.}

\item{Sex}{Number of the column containing sex 'M' for male and 'F' for
female, Default: 1}

\item{Pop}{Number of the column containing populations' names, Default: 2}

\item{firstX}{Number of column containing measured parameters (First of
multiple in case of multivariate analysis), Default: 3}

<<<<<<< HEAD
\item{test}{`1` for Greene t test [t_greene], `2` for [univariate], `3` for
sex specific ANOVA [aov_ss], and `4` for [multivariate], Default: 1}
=======
<<<<<<< HEAD
\item{test}{`1` for Greene t test [t_greene], `2` for [univariate], `3` for
sex specific ANOVA [aov_ss], and `4` for [multivariate], Default: 1}
=======
<<<<<<< HEAD
\item{test}{`1` for Greene t test [t_greene], `2` for [univariate], `3` for
sex specific ANOVA [aov_ss], and `4` for [multivariate], Default: 1}
=======
\item{test}{\code{1} for Greene t test \link{t_greene}, \code{2} for \link{univariate}, \code{3} for
sex specific ANOVA \link{aov_ss}, and \code{4} for \link{multivariate}, Default: 1}
>>>>>>> 6dea5f88f050c15e544b721a245e69cf92b48861
>>>>>>> 3ffea2d74b01d4945f77061b801bd9783c34951e
>>>>>>> 49bab1ae1b9ab0ccdc09b835509661bcaf96caeb

\item{run}{Logical; if TRUE runs the corresponding test after data
extraction, Default: TRUE}

\item{...}{Additional arguments that could be passed to the test of choice}
}
\value{
Input for other functions.
}
\description{
Extract summary data needed for other functions from raw data.
}
\details{
Raw data is entered in a wide format data frame similar to
<<<<<<< HEAD
[Howells] data set. The first two columns contain sex `Sex` (`M` for male
and `F` for female) (Default: `1`) and populations' names `Pop` (Default:
`2`). Starting from `firstX` column (Default: `3`), measured parameters are
=======
<<<<<<< HEAD
[Howells] data set. The first two columns contain sex `Sex` (`M` for male
and `F` for female) (Default: `1`) and populations' names `Pop` (Default:
`2`). Starting from `firstX` column (Default: `3`), measured parameters are
=======
<<<<<<< HEAD
[Howells] data set. The first two columns contain sex `Sex` (`M` for male
and `F` for female) (Default: `1`) and populations' names `Pop` (Default:
`2`). Starting from `firstX` column (Default: `3`), measured parameters are
=======
\link{Howells} data set. The first two columns contain sex \code{Sex} (\code{M} for male
and \code{F} for female) (Default: \code{1}) and populations' names \code{Pop} (Default:
\code{2}). Starting from \code{firstX} column (Default: \code{3}), measured parameters are
>>>>>>> 6dea5f88f050c15e544b721a245e69cf92b48861
>>>>>>> 3ffea2d74b01d4945f77061b801bd9783c34951e
>>>>>>> 49bab1ae1b9ab0ccdc09b835509661bcaf96caeb
entered each in a separate column.
}
\examples{
\donttest{
# for multivariate test
library(TestDimorph)
extract_sum(Howells, test = 4)
# for univariate test on a specific parameter
library(TestDimorph)
extract_sum(Howells, test = 2, firstX = 4)
}
}
