% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Howells}
\alias{Howells}
\title{The Howells' craniometric data}
\format{
A data frame with 441 rows and 10 variables:
 \describe{
\item{Sex}{'M' for male and 'F' for female}
\item{Pop}{Populations' names}
\item{GOL}{Glabello occipital length}
\item{NOL}{Nasio occipital length}
\item{BNL}{Bastion nasion length}
\item{BBH}{Basion bregma height}
\item{XCB}{Maximum cranial breadth}
\item{XFB}{Max frontal breadth}
\item{ZYB}{Bizygomatic breadth}
\item{AUB}{Biauricular breadth}
  }
}
\usage{
Howells
}
\description{
A subset of a dataset that consists of 82 craniometric measurements taken
from approximately two thousands and half human crania from 28
geographically diverse populations.
}
\references{
Howells WW (1989). *Skull shapes and the map: craniometric analyses in the dispersion of modern
Homo.*. Papers of the Peabody Museum of Archaeology and Ethnology, 79.
}
\keyword{datasets}
