% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Howells}
\alias{Howells}
\title{The Howells' craniometric data}
\format{
A data frame with 441 rows and 10 variables:
<<<<<<< HEAD
 \describe{
=======
<<<<<<< HEAD
 \describe{
=======
<<<<<<< HEAD
 \describe{
=======
\describe{
>>>>>>> 6dea5f88f050c15e544b721a245e69cf92b48861
>>>>>>> 3ffea2d74b01d4945f77061b801bd9783c34951e
>>>>>>> 49bab1ae1b9ab0ccdc09b835509661bcaf96caeb
\item{Sex}{'M' for male and 'F' for female}
\item{Pop}{Populations' names}
\item{GOL}{Glabello occipital length}
\item{NOL}{Nasio occipital length}
\item{BNL}{Bastion nasion length}
\item{BBH}{Basion bregma height}
\item{XCB}{Maximum cranial breadth}
\item{XFB}{Max frontal breadth}
\item{ZYB}{Bizygomatic breadth}
\item{AUB}{Biauricular breadth}
<<<<<<< HEAD
  }
=======
<<<<<<< HEAD
  }
=======
<<<<<<< HEAD
  }
=======
}
>>>>>>> 6dea5f88f050c15e544b721a245e69cf92b48861
>>>>>>> 3ffea2d74b01d4945f77061b801bd9783c34951e
>>>>>>> 49bab1ae1b9ab0ccdc09b835509661bcaf96caeb
}
\usage{
Howells
}
\description{
A subset of a dataset that consists of 82 craniometric measurements taken
from approximately two thousands and half human crania from 28
geographically diverse populations.
}
\references{
<<<<<<< HEAD
Howells WW (1989). *Skull shapes and the map: craniometric analyses in the dispersion of modern
Homo.*. Papers of the Peabody Museum of Archaeology and Ethnology, 79.
=======
<<<<<<< HEAD
Howells WW (1989). *Skull shapes and the map: craniometric analyses in the dispersion of modern
Homo.*. Papers of the Peabody Museum of Archaeology and Ethnology, 79.
=======
<<<<<<< HEAD
Howells WW (1989). *Skull shapes and the map: craniometric analyses in the dispersion of modern
Homo.*. Papers of the Peabody Museum of Archaeology and Ethnology, 79.
=======
Howells WW (1989). \emph{Skull shapes and the map: craniometric analyses in the dispersion of modern
Homo.}. Papers of the Peabody Museum of Archaeology and Ethnology, 79.
>>>>>>> 6dea5f88f050c15e544b721a245e69cf92b48861
>>>>>>> 3ffea2d74b01d4945f77061b801bd9783c34951e
>>>>>>> 49bab1ae1b9ab0ccdc09b835509661bcaf96caeb
}
\keyword{datasets}
