% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{multi_raw}
\alias{multi_raw}
\title{multivariate data generation}
\usage{
multi_raw(x, format, complete_cases, dist, lower, upper)
}
\arguments{
\item{x}{Data frame or list containing summary statistics for
multiple parameters measured in both sexes in two or more populations.}

\item{format}{form of the resultant tibble either 'long' or 'wide',
Default: 'wide'}

\item{complete_cases}{Logical; if TRUE rows with missing values will be
removed, Default: FALSE}

\item{dist}{univariate distribution used for data generation either `log`
for log normal or `truncated` for truncated distribution, Default: 'truncated'}

\item{lower}{vector of lower bounds, Default: -Inf}

\item{upper}{vector of upper bounds, Default: Inf}
}
\description{
multivariate data generation helper function
}
\keyword{internal}
