% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{multivariate}
\alias{multivariate}
\title{Multivariate Analysis Of Sexual Dimorphism}
\usage{
multivariate(
  x,
  R.res = NULL,
  Parms = 1,
  Pop = 2,
  es = FALSE,
  univariate = FALSE,
  padjust = p.adjust.methods,
  ...,
  lower.tail = FALSE,
  digits = 4
)
}
\arguments{
\item{x}{Tibble/Data frame or list containing summary statistics for
multiple parameters measured in both sexes in two or more populations.}

\item{R.res}{Pooled within correlational matrix, Default: NULL}

\item{Parms}{Number of the column containing names of measured parameters,
Default: 1}

\item{Pop}{Number of the column containing populations' names, Default: 2}

\item{es}{Logical; if TRUE effect size is included in the output , Default:
FALSE}

\item{univariate}{Logical; if TRUE conducts multiple univariate analyses on
different parameters separately, Default: FALSE}

\item{padjust}{Method of p.value adjustment for multiple comparisons
<<<<<<< HEAD
following [p.adjust.methods]}
=======
<<<<<<< HEAD
following [p.adjust.methods]}
=======
<<<<<<< HEAD
<<<<<<< HEAD
following [p.adjust.methods]}
=======
following \link{p.adjust.methods}}
>>>>>>> 6dea5f88f050c15e544b721a245e69cf92b48861
>>>>>>> 3ffea2d74b01d4945f77061b801bd9783c34951e
>>>>>>> 49bab1ae1b9ab0ccdc09b835509661bcaf96caeb
=======
following \link{p.adjust.methods}}
>>>>>>> 6dea5f88f050c15e544b721a245e69cf92b48861
>>>>>>> 3ffea2d74b01d4945f77061b801bd9783c34951e
>>>>>>> 49bab1ae1b9ab0ccdc09b835509661bcaf96caeb

\item{...}{Additional arguments that could be passed to the [univariate]
function}

\item{lower.tail}{Logical; if TRUE probabilities are `P[X <= x]`,
otherwise, `P[X > x]`., Default: FALSE}

\item{digits}{Number of significant digits, Default: 4}
}
\value{
Tibble of MANOVA results
}
\description{
<<<<<<< HEAD
Multivariate extension of Greene t test [t_greene]
=======
<<<<<<< HEAD
Multivariate extension of Greene t test [t_greene]
=======
<<<<<<< HEAD
<<<<<<< HEAD
Multivariate extension of Greene t test [t_greene]
=======
Multivariate extension of Greene t test \link{t_greene}
>>>>>>> 6dea5f88f050c15e544b721a245e69cf92b48861
>>>>>>> 3ffea2d74b01d4945f77061b801bd9783c34951e
>>>>>>> 49bab1ae1b9ab0ccdc09b835509661bcaf96caeb
=======
Multivariate extension of Greene t test \link{t_greene}
>>>>>>> 6dea5f88f050c15e544b721a245e69cf92b48861
>>>>>>> 3ffea2d74b01d4945f77061b801bd9783c34951e
>>>>>>> 49bab1ae1b9ab0ccdc09b835509661bcaf96caeb
}
\details{
Data can be entered either as a tibble/data frame of summary
statistics as in [baboon.parms_df] . In that case the pooled within
correlational matrix `R.res` should be entered as a separate argument as in
[R]. Another acceptable format is a named list of matrices containing
different summary statistics as well as the correlational matrix as in
[baboon.parms_list]. By setting the option `univariate` to `TRUE`, multiple
`ANOVA`s can be run on each parameter independently with the required p
value correction using [p.adjust.methods].
}
\examples{
\donttest{
# x is a data frame with separate correlational matrix
library(TestDimorph)
multivariate(baboon.parms_df, R.res = R)
# x is a list with the correlational matrix included
library(TestDimorph)
multivariate(baboon.parms_list, univariate = TRUE, padjust = "bonferroni")
}
}
