% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{padjust_n}
\alias{padjust_n}
\title{padjust_n}
\usage{
padjust_n(p, method = p.adjust.methods, n = length(p))
}
\arguments{
\item{p}{pvalue to be adjusted}

\item{method}{method of adjustment, Default: p.adjust.methods}

\item{n}{number of pairwise comparisons, Default: length(p)}
}
\description{
doesn't return an error when n >= lp
}
\keyword{internal}
