% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMatrix.R
\name{pMatrix-deprecated}
\alias{pMatrix-deprecated}
\title{Visualization Of t-Greene Pairwise Comparisons}
\value{
Graphical or numerical matrix of p values from Greene t test
pairwise comparisons. N.B: contrary to the usual corrplots where higher
values indicate stronger correlation, here lower values indicate
significance
}
\description{
Returns a graphical or numerical correlational matrix of p
values for the interpopulation degree of sexual dimorphism as measured by
Greene t test
}
\details{
Data is entered as a tibble/data frame of summary statistics where
the column containing population names is chosen by position (first by
default), other columns of summary data should have specific names (case
sensitive) similar to \link{baboon.parms_df}
}
\seealso{
\code{\link[=TestDimorph-deprecated]{TestDimorph-deprecated()}}
}
\keyword{internal}
