% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{t_test}
\alias{t_test}
\title{t-greene test}
\usage{
t_test(
  m,
  f,
  m2,
  f2,
  M.mu,
  F.mu,
  M.mu2,
  F.mu2,
  M.sdev,
  F.sdev,
  M.sdev2,
  F.sdev2,
  padjust = padjust,
  N,
  digits,
  sig.level,
  alternative,
  es
)
}
\arguments{
\item{m}{Number of male sample size in the first population, Default: NULL}

\item{f}{Number of female sample size in the first population, Default:
NULL}

\item{m2}{Number of male sample size in the second population, Default:
NULL}

\item{f2}{Number of female sample size in the second population, Default:
NULL}

\item{M.mu}{Means for males in the first population, Default: NULL}

\item{F.mu}{Means for females in the first population, Default: NULL}

\item{M.mu2}{Means for males in the second population, Default: NULL}

\item{F.mu2}{Means for females in the second population, Default: NULL}

\item{M.sdev}{Standard deviation for males in the first population,
Default: NULL}

\item{F.sdev}{Standard deviation for females in the first population,
Default: NULL}

\item{M.sdev2}{Standard deviation for males in the second population,
Default: NULL}

\item{F.sdev2}{Standard deviation for females in the second population,
Default: NULL}

\item{padjust}{Method of p.value adjustment for multiple comparisons
<<<<<<< HEAD
following [p.adjust.methods]}

\item{N}{Number of pairwise comparisons for [p.adjust.methods], if left
`NULL` it will follow the formula `n(n 1)/2` where `n` is the number of
=======
<<<<<<< HEAD
following [p.adjust.methods]}

\item{N}{Number of pairwise comparisons for [p.adjust.methods], if left
`NULL` it will follow the formula `n(n 1)/2` where `n` is the number of
=======
following \link{p.adjust.methods}}

\item{N}{Number of pairwise comparisons for \link{p.adjust.methods}, if left
\code{NULL} it will follow the formula \verb{n(n 1)/2} where \code{n} is the number of
>>>>>>> 6dea5f88f050c15e544b721a245e69cf92b48861
>>>>>>> 3ffea2d74b01d4945f77061b801bd9783c34951e
populations , Default: NULL}

\item{digits}{Number of significant digits, Default: 4}

\item{sig.level}{Critical p.value, Default: 0.05}

\item{alternative}{a character string specifying the alternative
hypothesis, must be one of "two.sided", "greater" or "less".}

\item{es}{Logical; if TRUE effect size is included in the output , Default:
FALSE}
}
\description{
t-greene test
}
\keyword{internal}
