% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestDimorph-deprecated.R, R/pMatrix.R
\name{TestDimorph-deprecated}
\alias{TestDimorph-deprecated}
\alias{pMatrix}
\title{Deprecated functions in package \pkg{TestDimorph}.}
\usage{
pMatrix(
  x,
  Pop = 1,
  plot = FALSE,
  padjust = p.adjust.methods,
  CI = 0.95,
  digits = 4,
  alternative = c("two.sided", "less", "greater"),
  ...
)
}
\description{
The functions listed below are deprecated and will be defunct
in the near future.
}
\section{`pMatrix`}{

For `pMatrix`, use [t_greene()]
}

\keyword{internal}
