% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccuModel.R, R/RawGen.R,
%   R/TestDimorph-deprecated.R, R/Tg.R, R/aovSS.R, R/pMatrix.R
\name{AccuModel}
\alias{AccuModel}
\alias{RawGen}
\alias{TestDimorph-deprecated}
\alias{Tg}
\alias{aovSS}
\alias{pMatrix}
\title{Deprecated functions in package \pkg{TestDimorph}.}
\usage{
AccuModel(
  f,
  x,
  y = NULL,
  method = "lda",
  res_method = "repeatedcv",
  p = 0.75,
  nf = 10,
  nr = 3,
  plot = FALSE,
  Sex = 1,
  Pop = NULL,
  byPop = FALSE,
  ref. = "F",
  post. = "M",
  ...
)

RawGen(
  x,
  Parms = 1,
  Pop = 2,
  R.res = NULL,
  dist = c("truncated", "log"),
  lower = -Inf,
  upper = Inf,
  format = c("wide", "long"),
  complete_cases = FALSE
)

Tg(
  x = NULL,
  Pop = 1,
  es = FALSE,
  plot = FALSE,
  ...,
  alternative = c("two.sided", "less", "greater"),
  padjust = p.adjust.methods,
  letters = FALSE,
  digits = 4,
  sig.level = 0.05
)

aovSS(
  x,
  Pop = 1,
  pairwise = TRUE,
  letters = FALSE,
  es = FALSE,
  digits = 4,
  sig.level = 0.05
)

pMatrix(
  x,
  Pop = 1,
  plot = FALSE,
  padjust = p.adjust.methods,
  sig.level = 0.05,
  digits = 4,
  alternative = c("two.sided", "less", "greater"),
  ...
)
}
\description{
The functions listed below are deprecated and will be defunct
in the near future.
}
\section{\code{AccuModel}}{

For \code{AccuModel}, use \code{\link[=accu_model]{accu_model()}}.
}

\section{\code{RawGen}}{

For \code{RawGen}, use \code{\link[=raw_gen]{raw_gen()}}.
}

\section{\code{Tg}}{

For \code{Tg}, use \code{\link[=t_greene]{t_greene()}}.
}

\section{\code{aovSS}}{

For \code{aovSS}, use \code{\link[=aov_ss]{aov_ss()}}
}

\section{\code{pMatrix}}{

For \code{pMatrix}, use \code{\link[=t_greene]{t_greene()}}
}

\keyword{internal}
