% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccuModel.R, R/RawGen.R,
%   R/TestDimorph-deprecated.R, R/Tg.R, R/aovSS.R, R/pMatrix.R
\name{AccuModel}
\alias{AccuModel}
\alias{RawGen}
\alias{TestDimorph-deprecated}
\alias{Tg}
\alias{aovSS}
\alias{pMatrix}
\title{Deprecated functions in package \pkg{TestDimorph}.}
\usage{
AccuModel(
  f,
  x,
  y = NULL,
  method = "lda",
  res_method = "repeatedcv",
  prior = NULL,
  p = 0.75,
  nf = 10,
  nr = 3,
  plot = FALSE,
  Sex = 1,
  Pop = NULL,
  byPop = FALSE,
  ref. = "F",
  post. = "M",
  ...
)

RawGen(
  x,
  Trait = 1,
  Pop = 2,
  R.res = NULL,
  dist = c("truncated", "log"),
  lower = -Inf,
  upper = Inf,
  format = c("wide", "long"),
  complete_cases = FALSE
)

Tg(
  x = NULL,
  Pop = 1,
  es = "none",
  plot = FALSE,
  ...,
  alternative = c("two.sided", "less", "greater"),
  padjust = "none",
  letters = FALSE,
  digits = 4,
  CI = 0.95
)

aovSS(
  x,
  Pop = 1,
  pairwise = TRUE,
  letters = FALSE,
  es_anova = "none",
  digits = 4,
  CI = 0.95
)

pMatrix(
  x,
  Pop = 1,
  plot = FALSE,
  padjust = p.adjust.methods,
  CI = 0.95,
  digits = 4,
  alternative = c("two.sided", "less", "greater"),
  ...
)
}
\description{
The functions listed below are deprecated and will be defunct
in the near future.
}
\section{`AccuModel`}{

For `AccuModel`, use [accu_model()].
}

\section{`RawGen`}{

For `RawGen`, use [raw_gen()].
}

\section{`Tg`}{

For `Tg`, use [t_greene()].
}

\section{`aovSS`}{

For `aovSS`, use [aov_ss()]
}

\section{`pMatrix`}{

For `pMatrix`, use [t_greene()]
}

\keyword{internal}
