% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccuModel.R, R/RawGen.R,
%   R/TestDimorph-deprecated.R, R/Tg.R, R/aovSS.R, R/pMatrix.R
\name{AccuModel}
\alias{AccuModel}
\alias{RawGen}
\alias{TestDimorph-deprecated}
\alias{Tg}
\alias{aovSS}
\alias{pMatrix}
\title{Deprecated functions in package \pkg{TestDimorph}.}
\usage{
AccuModel(
  f,
  x,
  y = NULL,
  method = "lda",
  res_method = "repeatedcv",
  p = 0.75,
  nf = 10,
  nr = 3,
  plot = FALSE,
  Sex = 1,
  Pop = NULL,
  byPop = FALSE,
  ref. = "F",
  post. = "M",
  ...
)

RawGen(
  x,
  Parms = 1,
  Pop = 2,
  R.res = NULL,
  dist = c("truncated", "log"),
  lower = -Inf,
  upper = Inf,
  format = c("wide", "long"),
  complete_cases = FALSE
)

Tg(
  x = NULL,
  Pop = 1,
  es = FALSE,
  plot = FALSE,
  ...,
  alternative = c("two.sided", "less", "greater"),
  padjust = p.adjust.methods,
  letters = FALSE,
  digits = 4,
  sig.level = 0.05
)

aovSS(
  x,
  Pop = 1,
  pairwise = TRUE,
  letters = FALSE,
  es = FALSE,
  digits = 4,
  sig.level = 0.05
)

pMatrix(
  x,
  Pop = 1,
  plot = FALSE,
  padjust = p.adjust.methods,
  sig.level = 0.05,
  digits = 4,
  alternative = c("two.sided", "less", "greater"),
  ...
)
}
\description{
The functions listed below are deprecated and will be defunct
in the near future.
}
<<<<<<< HEAD
\section{`AccuModel`}{
=======
<<<<<<< HEAD
\section{`AccuModel`}{
<<<<<<< HEAD
=======
<<<<<<< HEAD
\section{`AccuModel`}{
=======
>>>>>>> 49bab1ae1b9ab0ccdc09b835509661bcaf96caeb

For `AccuModel`, use [accu_model()].
}

\section{`RawGen`}{

For `RawGen`, use [raw_gen()].
}

\section{`Tg`}{

For `Tg`, use [t_greene()].
}

\section{`aovSS`}{

For `aovSS`, use [aov_ss()]
}

\section{`pMatrix`}{

For `pMatrix`, use [t_greene()]
=======
\section{\code{AccuModel}}{
>>>>>>> 3ffea2d74b01d4945f77061b801bd9783c34951e
<<<<<<< HEAD
>>>>>>> 49bab1ae1b9ab0ccdc09b835509661bcaf96caeb
=======
>>>>>>> 49bab1ae1b9ab0ccdc09b835509661bcaf96caeb

For `AccuModel`, use [accu_model()].
}

\section{`RawGen`}{

For `RawGen`, use [raw_gen()].
}

\section{`Tg`}{

For `Tg`, use [t_greene()].
}

\section{`aovSS`}{

For `aovSS`, use [aov_ss()]
}

\section{`pMatrix`}{

<<<<<<< HEAD
For `pMatrix`, use [t_greene()]
=======
<<<<<<< HEAD
<<<<<<< HEAD
For `pMatrix`, use [t_greene()]
=======
For \code{pMatrix}, use \code{\link[=t_greene]{t_greene()}}
>>>>>>> 6dea5f88f050c15e544b721a245e69cf92b48861
>>>>>>> 3ffea2d74b01d4945f77061b801bd9783c34951e
>>>>>>> 49bab1ae1b9ab0ccdc09b835509661bcaf96caeb
=======
For \code{pMatrix}, use \code{\link[=t_greene]{t_greene()}}
>>>>>>> 6dea5f88f050c15e544b721a245e69cf92b48861
>>>>>>> 3ffea2d74b01d4945f77061b801bd9783c34951e
>>>>>>> 49bab1ae1b9ab0ccdc09b835509661bcaf96caeb
}

\keyword{internal}
