% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate}
\alias{univariate}
\title{Univariate Analysis Of Sexual Dimorphism}
\usage{
univariate(
  x,
  Pop = 1,
  es = FALSE,
  pairwise = FALSE,
  padjust = p.adjust.methods,
  ...,
  lower.tail = FALSE,
  N = NULL,
  digits = 4
)
}
\arguments{
\item{x}{A data frame containing summary statistics.}

\item{Pop}{Number of the column containing populations' names, Default: 1}

\item{es}{Logical; if TRUE effect size is included in the output , Default:
FALSE}

\item{pairwise}{Logical; if TRUE runs multiple pairwise comparisons on
<<<<<<< HEAD
different populations using [t_greene] test, Default: FALSE}
=======
<<<<<<< HEAD
different populations using [t_greene] test, Default: FALSE}

\item{padjust}{Method of p.value adjustment for multiple comparisons
following [p.adjust.methods]}

\item{...}{Additional arguments that could be passed to the [t_greene]
=======
different populations using \link{t_greene} test, Default: FALSE}
>>>>>>> 3ffea2d74b01d4945f77061b801bd9783c34951e

\item{padjust}{Method of p.value adjustment for multiple comparisons
following [p.adjust.methods]}

<<<<<<< HEAD
\item{...}{Additional arguments that could be passed to the [t_greene]
=======
\item{...}{Additional arguments that could be passed to the \link{t_greene}
>>>>>>> 6dea5f88f050c15e544b721a245e69cf92b48861
>>>>>>> 3ffea2d74b01d4945f77061b801bd9783c34951e
function}

\item{lower.tail}{Logical; if TRUE probabilities are `P[X <= x]`,
otherwise, `P[X > x]`., Default: FALSE}

<<<<<<< HEAD
\item{N}{Number of pairwise comparisons for [p.adjust.methods], if left
`NULL` it will follow the formula `n(n 1)/2` where `n` is the number of
=======
<<<<<<< HEAD
\item{N}{Number of pairwise comparisons for [p.adjust.methods], if left
`NULL` it will follow the formula `n(n 1)/2` where `n` is the number of
=======
\item{N}{Number of pairwise comparisons for \link{p.adjust.methods}, if left
\code{NULL} it will follow the formula \verb{n(n 1)/2} where \code{n} is the number of
>>>>>>> 6dea5f88f050c15e544b721a245e69cf92b48861
>>>>>>> 3ffea2d74b01d4945f77061b801bd9783c34951e
populations , Default: NULL}

\item{digits}{Number of significant digits, Default: 4}
}
\value{
Tibble of ANOVA results
}
\description{
Calculation and visualization of the differences in degree
sexual dimorphism between multiple populations using a modified one way
ANOVA and summary statistics as input
}
\details{
Data is entered as a tibble/data frame of summary statistics where
the column containing population names is chosen by position (first by
default), other columns of summary data should have specific names (case
sensitive) similar to [baboon.parms_df]
}
\examples{
\donttest{
# Comparisons of femur head diameter in four populations
library(TestDimorph)
df <-
  data.frame(
    Pop = c("Turkish", "Bulgarian", "Greek", "Portuguese "),
    m = c(150.00, 82.00, 36.00, 34.00),
    M.mu = c(49.39, 48.33, 46.99, 45.20),
    M.sdev = c(3.01, 2.53, 2.47, 2.00),
    f = c(150.00, 58.00, 34.00, 24.00),
    F.mu = c(42.91, 42.89, 42.44, 40.90),
    F.sdev = c(2.90, 2.84, 2.26, 2.90)
  )
univariate(df, pairwise = TRUE, padjust = "bonferroni")
}
}
